修正：ログインstate、handleLogin、ログインボタンを一ヶ所にまとめたカスタムフックを作成
